refactor: kako-jun/name-name#36 EventDisplayをEventCardから分離
EventCard.tsx が 407 行あり、編集 UI と読み取り表示の責務が混ざっていた。
EventDisplay（variant ごとの読み取り専用表示）を EventDisplay.tsx に切り出し、EventCard は編集フォームとドロップ領域に集中させる。
行数: EventCard 407 → 277, EventDisplay 新規 141 行。
